requirements.txt
asgiref==3.5.0
Django==4.0.3
sqlparse==0.4.2
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2